pyproject.toml
[project]
name = "syncify2"
version = "0.0.0"
requires-python = ">= 3.13"
dependencies = ["flask", "spotipy", "flask_session2", "python-dotenv", "uvicorn", "sqlalchemy", "alembic"]

[tool.setuptools]
py-modules = ["webapp", "worker", "common"]
